+++ frontend/src/views/order/NoticeCard.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script setup lang='ts'>
-import List from './List.vue';
+import List from '../../components/order/List.vue';
 </script>
 
 <style scoped>
