+++ frontend/src/views/personal/Header.vue
@@ -22,7 +22,7 @@
 import { UserOutlined } from "@ant-design/icons-vue";
 import { getCurrentInstance, onMounted, ref } from "vue";
 import router from "@/router";
-import Avatar from "../order/Avatar.vue";
+import Avatar from "../../components/order/Avatar.vue";
 import Calander from "./Calander.vue";
 import { message } from "ant-design-vue";
 const uid = localStorage.getItem('uid')
